+++ AuroraDialogEnhancer.Extensions/Dimensions/Rgba.cs
@@ -1,15 +1,15 @@
 ﻿namespace AuroraDialogEnhancerExtensions.Dimensions;
 public class Rgba
 {
-    public int R { get; set; }
+    public byte R { get; set; }
 
-    public int G { get; set; }
+    public byte G { get; set; }
 
-    public int B { get; set; }
+    public byte B { get; set; }
 
-    public int A { get; set; }
+    public byte A { get; set; }
 
-    public Rgba(int a, int r, int g, int b)
+    public Rgba(byte a, byte r, byte g, byte b)
     {
         R = r;
         G = g;
